Added @table @columns. Moved db to model package
