+++ src/main/java/pl/java/spring/web/project/clientappproject/database/model/Client.java
@@ -1,31 +1,45 @@
-package pl.java.spring.web.project.clientappproject.database;
+package pl.java.spring.web.project.clientappproject.database.model;
 
-import javax.persistence.*;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.OneToMany;
+import javax.persistence.Table;
+
 import java.util.ArrayList;
 import java.util.Collection;
 
 @Entity
-public class Employee {
+@Table(name = "CLIENT")
+public class Client {
 
     @Id
+    @Column(name = "client_id")
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private long employeeId;
+    private long clientId;
 
+    @Column(name = "first_name")
     private String firstName;
+
+    @Column(name = "last_name")
     private String lastName;
 
     @OneToMany
     private Collection<Address> addresses = new ArrayList<>();
 
-    @ManyToMany
+    @OneToMany
     private Collection<Task> tasks = new ArrayList<>();
 
-    public long getEmployeeId() {
-        return employeeId;
+
+
+    public long getClientId() {
+        return clientId;
     }
 
-    public void setEmployeeId(long employeeId) {
-        this.employeeId = employeeId;
+    public void setClientId(long clientId) {
+        this.clientId = clientId;
     }
 
     public String getFirstName() {
