+++ src/main/java/pl/java/spring/web/project/clientappproject/database/model/Employee.java
@@ -1,31 +1,48 @@
-package pl.java.spring.web.project.clientappproject.database;
+package pl.java.spring.web.project.clientappproject.database.model;
 
-import javax.persistence.*;
+import org.springframework.stereotype.Controller;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.ManyToMany;
+import javax.persistence.OneToMany;
+import javax.persistence.Table;
+
 import java.util.ArrayList;
 import java.util.Collection;
 
 @Entity
-public class Client {
+@Table(name = "EMPLOYEE")
+public class Employee {
 
     @Id
+    @Column(name = "employee_id")
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private long clientId;
+    private long employeeId;
 
+    @Column(name = "first_name")
     private String firstName;
+
+    @Column(name = "last_name")
     private String lastName;
 
     @OneToMany
     private Collection<Address> addresses = new ArrayList<>();
 
-    @OneToMany
+    @ManyToMany
     private Collection<Task> tasks = new ArrayList<>();
 
-    public long getClientId() {
-        return clientId;
+
+
+    public long getEmployeeId() {
+        return employeeId;
     }
 
-    public void setClientId(long clientId) {
-        this.clientId = clientId;
+    public void setEmployeeId(long employeeId) {
+        this.employeeId = employeeId;
     }
 
     public String getFirstName() {
